Rename Lab3.py to Student Class.py
You will be creating two classes: 1) Person Class: o This Class has three private attributes: name, email, and age. o The constructor will take three attributes, as mentioned in the diagram, in order to update the content of the attributes. Regarding the age, the constructor will be calling internally calculateAge() function with the birthday value, this function will update the age.  Hint: use the datetime module for calculateAge() o Define the setters and getters of the variables as mentioned in the diagram. o In set_email(email) make sure that it is proper email (has one @ followed by period not directly) o Override __str__() function so that it will return all the values of the attributes of a Person in a proper format. o Override __eq__() function so it will return true if two Person instances has the same name.  This will enable us to compare two Person instances using == 2) Student Class:  This class inherits from Person class, so it will have in addition to the three inherited data variables, other three private variables as follows:  SID: Student ID  courses: to hold the courses along the grades of this student; it will be created as dictionary where the key is the course ID (for example CS120) and the value will be the grade of that course.  AverageGrade: that will be calculated in your class.  It will have extra functions, in addition to the inherited ones, as follows;  The constructor that will update the values of the variables, and also calls the super constructor.  addCourse(CourseID: String, result: int): that will take the key and the grade to add a single course at a time to the courses variable.  calculateAverage(): will not take any parameter, it will calculate and return the average of the grades in the courses for a given student.  Define the setters and getters of the variables as mentioned in the diagram.  Override __str__() function so that it will return all the values of the attributes of a Student in a proper format. (don’t forget to use the __str__() of the parent class as well)  Override as well the __lt__() function that will compare two Student Objects according to the averageGrade. (This will help us to use the List.sort() simply with a list of Students to sort them automatically according to averageGrade)